Added group - student connection Refactored into apis application
