+++ apis/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+from . import views
+
+urlpatterns=[
+    path('assessments/',views.assessments),
+    path('assessments/<int:id>',views.assessment),
+    path('groups/',views.groups),
+    path('groups/<int:id>',views.group),
+    path('students/',views.students),
+    path('students/<int:id>',views.student),
+]
