--- assessments/serializers.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from .models import Assessment
-
-class AssessmentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Assessment
-        fields = ['id', 'name', 'date']
