--- assessments/views.py
@@ -1,42 +1,0 @@
-from django.http import JsonResponse
-from .models import Assessment
-from .serializers import AssessmentSerializer
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from rest_framework import status
-
-@api_view(['GET', 'POST'])
-def assessments(request):
-
-    if request.method == 'GET':
-        assessments = Assessment.objects.all()
-        serializer = AssessmentSerializer(assessments, many=True)
-        return Response(serializer.data)
-
-    elif request.method == 'POST':
-        serializer = AssessmentSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status = status.HTTP_201_CREATED)
-        return Response(status.HTTP_400_BAD_REQUEST)
-
-@api_view(['GET', 'PUT', 'DELETE'])
-def assessment(request, id):
-
-    try:
-        assessment = Assessment.objects.get(pk=id)
-    except Assessment.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-
-    if request.method == 'GET':
-        serializer = AssessmentSerializer(assessment)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = AssessmentSerializer(assessment, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(status = status.HTTP_400_BAD_REQUEST)
-    elif request.method == 'DELETE':
-        assessment.delete()
-        return Response(status = status.HTTP_204_NO_CONTENT)
