--- groups/serializers.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from .models import Group
-
-class GroupSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Group
-        fields = ['id', 'name', 'date']
