--- groups/views.py
@@ -1,42 +1,0 @@
-from django.http import JsonResponse
-from .models import Group
-from .serializers import GroupSerializer
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from rest_framework import status
-
-@api_view(['GET', 'POST'])
-def groups(request):
-
-    if request.method == 'GET':
-        groups = Group.objects.all()
-        serializer = GroupSerializer(groups, many=True)
-        return Response(serializer.data)
-
-    elif request.method == 'POST':
-        serializer = GroupSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status = status.HTTP_201_CREATED)
-        return Response(status.HTTP_400_BAD_REQUEST)
-
-@api_view(['GET', 'PUT', 'DELETE'])
-def group(request, id):
-
-    try:
-        group = Group.objects.get(pk=id)
-    except Group.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-
-    if request.method == 'GET':
-        serializer = GroupSerializer(group)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = GroupSerializer(group, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(status = status.HTTP_400_BAD_REQUEST)
-    elif request.method == 'DELETE':
-        group.delete()
-        return Response(status = status.HTTP_204_NO_CONTENT)
