+++ quickcheck/settings.py
@@ -38,9 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework',
-    'students',
-    'groups',
-    'assessments',
+    'apis',
 ]
 
 MIDDLEWARE = [
