--- students/models.py
@@ -1,8 +1,0 @@
-from django.db import models
-
-class Student(models.Model):
-    name = models.CharField(max_length=200)
-    date = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return self.name
