--- students/serializers.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from .models import Student
-
-class StudentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Student
-        fields = ['id', 'name', 'date']
