--- students/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns=[
-    path('',views.students),
-    path('<int:id>',views.student),
-]
