--- students/views.py
@@ -1,42 +1,0 @@
-from django.http import JsonResponse
-from .models import Student
-from .serializers import StudentSerializer
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from rest_framework import status
-
-@api_view(['GET', 'POST'])
-def students(request):
-
-    if request.method == 'GET':
-        students = Student.objects.all()
-        serializer = StudentSerializer(students, many=True)
-        return Response(serializer.data)
-
-    elif request.method == 'POST':
-        serializer = StudentSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status = status.HTTP_201_CREATED)
-        return Response(status.HTTP_400_BAD_REQUEST)
-
-@api_view(['GET', 'PUT', 'DELETE'])
-def student(request, id):
-
-    try:
-        student = Student.objects.get(pk=id)
-    except Student.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-
-    if request.method == 'GET':
-        serializer = StudentSerializer(student)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = StudentSerializer(student, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(status = status.HTTP_400_BAD_REQUEST)
-    elif request.method == 'DELETE':
-        student.delete()
-        return Response(status = status.HTTP_204_NO_CONTENT)
